Unify config
The configuration files for the community portal are now loaded and
handled in one place. Before that every module opened and parsed the
config file on its own. Now all processing is done in the config module.
The rest of the code can simply access the properties of the config
object.