Drop captcha key_location and auto-create key file
The config option for captcha key_location is now replaced by
db_directory with a fixed file name 'captcha.key' in order to simplify
the configuration even further. The directory and key file are created
on startup if they don't exist yet. That makes deployment and testing
easier, too.

Further more I have tighten security with a couple of modes / chmod and
a secure umask.